// fore-end-Learning-log/vue3-ts-cms/src/store/main/system/type.ts
export interface IUser{
  id: number
  name: string
  realname: string
  cellphone: number
  enable: number
  departmentId: number
  roleId: number
  createAt: string
  updateAt: string
}

export interface ISystemState{
  usersList: IUser[]
  usersTotalCount: number

  pageList: any[]
  pageTotalCount: number
}